+++ indexCommon/src/main/java/ar/uba/fi/tppro/core/index/versionTracker/LocalShardVersionTracker.java
@@ -1,36 +1,29 @@
 package ar.uba.fi.tppro.core.index.versionTracker;
 
 import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.google.common.collect.HashMultimap;
 import com.google.common.collect.LinkedListMultimap;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Multimap;
-import com.google.common.collect.Multimaps;
-import com.netflix.curator.framework.CuratorFramework;
-import com.netflix.curator.framework.recipes.shared.SharedCount;
-import com.netflix.curator.framework.recipes.shared.SharedCountListener;
-import com.netflix.curator.framework.recipes.shared.SharedCountReader;
-import com.netflix.curator.framework.state.ConnectionState;
 
 public class LocalShardVersionTracker implements ShardVersionTracker {
 	
 	final Logger logger = LoggerFactory.getLogger(LocalShardVersionTracker.class);
 
-	private ConcurrentMap<Integer, AtomicInteger> versions = Maps.newConcurrentMap();
+	private ConcurrentMap<Integer, AtomicLong> versions = Maps.newConcurrentMap();
 	private Multimap<Integer, ShardVersionObserver> observers = LinkedListMultimap.create();
 
 	@Override
-	public int getCurrentVersion(int shardId)
+	public long getCurrentVersion(int shardId)
 			throws VersionTrackerServerException {
 		
-		AtomicInteger version = versions.get(shardId);
+		AtomicLong version = versions.get(shardId);
 		if(version == null){
-			version = new AtomicInteger(0);
+			version = new AtomicLong(0);
 			versions.put(shardId, version);
 		}
 		
@@ -39,12 +32,12 @@ public class LocalShardVersionTracker implements ShardVersionTracker {
 	}
 
 	@Override
-	public void setShardVersion(int shardId, int newVersion)
+	public void setShardVersion(int shardId, long newVersion)
 			throws StaleVersionException, VersionTrackerServerException {
 		
-		AtomicInteger version = versions.get(shardId);
+		AtomicLong version = versions.get(shardId);
 		if(version == null){
-			version = new AtomicInteger(0);
+			version = new AtomicLong(0);
 			versions.put(shardId, version);
 		}
