+++ indexCommon/src/main/java/ar/uba/fi/tppro/core/index/versionTracker/StaleVersionException.java
@@ -2,7 +2,7 @@ package ar.uba.fi.tppro.core.index.versionTracker;
 
 public class StaleVersionException extends Exception {
 
-	public StaleVersionException(int count) {
+	public StaleVersionException(long count) {
 		super("Different version in the server: " + count );
 	}
