+++ indexCommon/src/main/resources/IndexNode.thrift
@@ -75,7 +75,7 @@ service IndexBroker{
 service IndexNode{
 	QueryResult search(1: i32 shardId, 2: i32 partitionId, 3: string query, 4: i32 limit, 5: i32 offset) throws (1:ParseException parsex, 2:NonExistentPartitionException partex),
 	void deleteByQuery(1: i32 shardId, 2: i32 partitionId, 3: string query),
-	void prepareCommit(1: i32 shardId, 2: i32 partitionId, 3: i32 messageId, 4: list<Document> documents) throws (1: NonExistentPartitionException nonEx, 2: IndexException indexEx),
+	void prepareCommit(1: i32 shardId, 2: i32 partitionId, 3: i64 messageId, 4: list<Document> documents) throws (1: NonExistentPartitionException nonEx, 2: IndexException indexEx),
 	void commit(1: i32 shardId, 2: i32 partitionId) throws (1: NonExistentPartitionException nonEx, 2: IndexException indexEx),
 	void createPartition(1: i32 shardId, 2: i32 partitionId) throws (1: PartitionAlreadyExistsException partex),
 	void removePartition(1: i32 shardId, 2: i32 partitionId) throws (1: NonExistentPartitionException partex),
